+++ apps/tan-react/src/globals/components/Footer.tsx
@@ -1,6 +1,6 @@
 import { MyGithub } from './MyGithub';
 
-export default function Footer() {
+export const Footer = () => {
   const year = new Date().getFullYear();
 
   return (
@@ -13,4 +13,4 @@ export default function Footer() {
       </div>
     </footer>
   );
-}
+};