+++ apps/tan-react/src/globals/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './test.hooks';
+export * from './useTheme';
 export * from './wrapper';