+++ apps/tan-react/vitest.config.ts
@@ -1,13 +1,12 @@
-import { shared } from '@bemedev/shared-vitest';
+import { defineProject } from '@bemedev/dev-utils/vitest-extended';
 import viteReact from '@vitejs/plugin-react';
 
-export default shared({
+export default defineProject({
   plugins: [viteReact()],
 
   test: {
     name: 'tan-react',
     environment: 'jsdom',
-    passWithNoTests: true,
     globals: true,
     env: { RTL_SKIP_AUTO_CLEANUP: 'true' },
   },