+++ packages/__tests__/project1/vitest.config.ts
@@ -1,8 +1,8 @@
 import { exclude } from '@bemedev/dev-utils/vitest-exclude';
-import { shared } from '@bemedev/shared-vitest';
+import { defineProject } from '@bemedev/dev-utils/vitest-extended';
 import viteReact from '@vitejs/plugin-react';
 
-export default shared({
+export default defineProject({
   plugins: [viteReact(), exclude({ ignoreCoverageFiles: ['**/*.js'] })],
   test: {
     name: 'project1',