+++ packages/utils/vitest/rolldown.config.ts
@@ -1,3 +1,0 @@
-import { defineConfig } from '@bemedev/dev-utils/rolldown';
-
-export default defineConfig.bemedev();